major revision

+++ ScienceFair2015PC_Client/src/org/scifair2015/commands/Command.java
@@ -4,7 +4,7 @@ import java.io.DataOutputStream;
 
 public abstract class Command {
 
-	protected byte cmd;
+	protected int cmd;
 	
 	public int getCmd()
 	{
